move more tests to aave module + make things private
